Remove unnecessary files.
