+++ lib/src/web/factory_impl.dart
@@ -50,3 +50,24 @@ class RTCFactoryWeb extends RTCFactory {
   @override
   Navigator get navigator => NavigatorWeb();
 }
+
+Future<RTCPeerConnection> createPeerConnection(
+    Map<String, dynamic> configuration,
+    [Map<String, dynamic>? constraints]) {
+  return RTCFactoryWeb.instance
+      .createPeerConnection(configuration, constraints);
+}
+
+Future<MediaStream> createLocalMediaStream(String label) {
+  return RTCFactoryWeb.instance.createLocalMediaStream(label);
+}
+
+MediaRecorder mediaRecorder() {
+  return RTCFactoryWeb.instance.mediaRecorder();
+}
+
+VideoRenderer videoRenderer() {
+  return RTCFactoryWeb.instance.videoRenderer();
+}
+
+Navigator get navigator => RTCFactoryWeb.instance.navigator;
