--- lib/src/web/rtc_peerconnection_factory.dart
@@ -1,26 +1,0 @@
-import 'dart:async';
-
-import 'package:webrtc_interface/webrtc_interface.dart';
-
-import 'factory_impl.dart';
-
-Future<RTCPeerConnection> createPeerConnection(
-    Map<String, dynamic> configuration,
-    [Map<String, dynamic>? constraints]) {
-  return RTCFactoryWeb.instance
-      .createPeerConnection(configuration, constraints);
-}
-
-Future<MediaStream> createLocalMediaStream(String label) {
-  return RTCFactoryWeb.instance.createLocalMediaStream(label);
-}
-
-MediaRecorder mediaRecorder() {
-  return RTCFactoryWeb.instance.mediaRecorder();
-}
-
-VideoRenderer videoRenderer() {
-  return RTCFactoryWeb.instance.videoRenderer();
-}
-
-Navigator get navigator => RTCFactoryWeb.instance.navigator;
